[tcs toup] drivers: display: ili9xxx: configurable transmit function
This is a preparation for implementing potentially different
algorithms for the transfer of register or memory content, e.g.
to support not only SPI interfaces or very rarely encountered
hardware serial to parallel shift registers. For this purpose,
the previously globally accessible transfer function is moved
to the configuration area of the driver as a function pointer
and at the same time the distinction between access to
registers or display memory is introduced.

The previous implementation of the SPI transfer function is
unchanged and the behavior remains the same.

Signed-off-by: Stephan Linz <linz@li-pro.net>