[tcs toup] drivers: display: ili9xxx: add 16-bit shifted SPI to MCU transfer
In rare types of displays, discrete shift registers can be found
whose task is to convert the serial SPI bit stream to the parallel
MCU interface with clock and bit accuracy. Typically, these are
16 bits wide. However, wider expansion stages are also conceivable.

Examples of this type of display are the Pico-ResTouch-LCD-3.5
shield and the Pico Eval Board, both from Waveshare. They are
based on the ILI9488, which is not connected directly via SPI,
but via a discrete 16-bit deep shift register made up of the
chips 74HC4094 and 74HC4040. For more detials, read the
schematic as linked on the original Wiki page (Documents):

- https://www.waveshare.com/wiki/Pico-ResTouch-LCD-3.5
- https://www.waveshare.com/wiki/Pico-Eval-Board

Signed-off-by: Stephan Linz <linz@li-pro.net>