[tcs temphack] drivers: pwm: rpi_pico: avoid implicit float type cast
GCC Warning was: implicit conversion from 'float' to 'double'
to match other operand of binary expression [-Wdouble-promotion]

This is because of '-fsingle-precision-constant' which prohibits
implicit conversion of floats to doubles, is not set in CFLAGS per
default and so the ifloating point constant 16.0 will converted implicitly
to type 'double'. We will or can not use double precision here
and use the constant as float explicitly.

Signed-off-by: Stephan Linz <linz@li-pro.net>