[tcs toup] tests: kernel/cache: skip d-cache invalidation if Arm CMx
Blind and random invalidation of the data cache will tend to trash any
pending stack writes, among other things, and that's hard to unwind and
will cause unpredictable behaviour. Use of the L1 cache function
'SCB_InvalidateDCache()' from the CMSIS v5 Core API for Cortex-M usually
tends to endless loops or, in rare cases, to BUS or HARD FAULTS.

The Zephyr test "test_data_cache_api" with the call of
'sys_cache_data_invd_all()' will usually fail with a timeout on a
Cortex-M, rarely crash, because the corresponding CMSIS functions for
cache maintenance (invalidation) are not always reliable. Random use of
SCB_InvalidateDCache() has a lot of side-effects, often manifesting in
stack corruption. The hazard here is that the pending write-back or
write-buffers get lost, so the data visible in memory is not reflective
of what was written last (most recently). The pending data could be the
entire cache-line, and thus all the pushed registers.

Signed-off-by: Stephan Linz <linz@li-pro.net>