[tcs noup] doc: scripts: boards: skip strange boards
With the introduction of the board catalog by commit f2f149615601,
all boards, including those provided by Zephyr modules other than
Zephyr, are found at any time and attempts are made to include
them in the catalog. However, these “external” boards in particular
do not pass the necessary relation check to the Zephyr project and
the build-up of the catalog fails; same for undocumented boards.

This HOT fix tolerates the existence of other (foreign) boards and
simply skips them. This means that they are not included in the
board catalog as intended.

Signed-off-by: Stephan Linz <linz@li-pro.net>
Signed-off-by: Tobias Kaestner <tobias.kaestner@inovex.de>