[tcs noup] doc: scripts: boards: skip strange shields
With the expansions of the board catalog by commit 79e3ecb7a9e6,
all shields, including those provided by Zephyr modules other than
Zephyr, are found at any time and attempts are made to include
them in the catalog. However, these “external” shields in particular
do not pass the necessary relation check to the Zephyr project and
the build-up of the catalog fails; same for undocumented boards.

This HOT fix tolerates the existence of other (foreign) shields and
simply skips them. This means that they are not included in the
board & shield catalog as intended.

Signed-off-by: Stephan Linz <linz@li-pro.net>