[tcs toup] tests: spi_loopback: skip unsupported wordsize tests
Commit 60d9a16040e6bca17cb0e117590e91ea19d26e0b intrduced new tests
for different data frame sizes (7, 9, 16, 24, 32, 64 bit word size),
as long as the controller can handle that word size.

The skipping behavior of ZTest was not clear and a simple return
was used. But that seams to lead to more false positive failures
and a simple ztest skip should be the right solution to handle
the unsupported word sizes.

Signed-off-by: Stephan Linz <linz@li-pro.net>