[tcs noup 97437] boards: shields: tcan4550evm: support for FRDM-MCXA153
The board-specific device tree overlay for FRDM-MCXA153
has been added, which should be used, among other things,
due to reset issues with hardware-controlled CS pins.
Moreover, the second CS1 is used on this board at the
Arduino connector. This means that the CAN controller
node must be completely rebuilt with the reg index 1.

Signed-off-by: Stephan Linz <linz@li-pro.net>