[tcs toup] boards: shields: mikroe_can_fd_6_click: support for FRDM-RW612
The board-specific device tree overlay for FRDM-RW612 has been added,
which should be used due to reset issues with GPIO-controlled CS pins.

Moreover, the second CS1 is used on this board at the MicroBus connector.
This means that the CAN controller node must be completely rebuilt with
the reg index 1.

Signed-off-by: Stephan Linz <linz@li-pro.net>